add EtlSession class
