+++ stacks/data/pyspark/storage_utils.py
@@ -2,67 +2,14 @@
 
 import json
 import logging
-import os
 
 from azure.storage.blob import BlobServiceClient
 from azure.identity import DefaultAzureCredential
-from pyspark.sql import SparkSession
 
-from stacks.data.constants import (
-    AZURE_TENANT_ID,
-    AZURE_CLIENT_ID,
-    AZURE_CLIENT_SECRET,
-    AZURE_STORAGE_ACCOUNT_NAME,
-    AZURE_CONFIG_ACCOUNT_NAME,
-)
+from stacks.data.constants import AZURE_CONFIG_ACCOUNT_NAME
 
 
 logger = logging.getLogger(__name__)
-
-
-def check_env() -> None:
-    """Checks if the environment variables for ADLS and Blob access are set.
-
-    Raises:
-        EnvironmentError: If any of the required environment variables are not set.
-    """
-    required_variables = [
-        "AZURE_TENANT_ID",
-        "AZURE_CLIENT_ID",
-        "AZURE_CLIENT_SECRET",
-        "AZURE_STORAGE_ACCOUNT_NAME",
-        "AZURE_CONFIG_ACCOUNT_NAME",
-    ]
-
-    missing_variables = [var_name for var_name in required_variables if os.environ.get(var_name) is None]
-
-    if missing_variables:
-        raise EnvironmentError("The following environment variables are not set: " + ", ".join(missing_variables))
-
-
-def set_spark_properties(spark: SparkSession) -> None:
-    """Sets Spark properties to configure Azure credentials to access Data Lake storage.
-
-    Args:
-        spark: Spark session.
-    """
-    spark.conf.set(f"fs.azure.account.auth.type.{AZURE_STORAGE_ACCOUNT_NAME}.dfs.core.windows.net", "OAuth")
-    spark.conf.set(
-        f"fs.azure.account.oauth.provider.type.{AZURE_STORAGE_ACCOUNT_NAME}.dfs.core.windows.net",
-        "org.apache.hadoop.fs.azurebfs.oauth2.ClientCredsTokenProvider",
-    )
-    spark.conf.set(
-        f"fs.azure.account.oauth2.client.id.{AZURE_STORAGE_ACCOUNT_NAME}.dfs.core.windows.net",
-        AZURE_CLIENT_ID,
-    )
-    spark.conf.set(
-        f"fs.azure.account.oauth2.client.secret.{AZURE_STORAGE_ACCOUNT_NAME}.dfs.core.windows.net",
-        AZURE_CLIENT_SECRET,
-    )
-    spark.conf.set(
-        f"fs.azure.account.oauth2.client.endpoint.{AZURE_STORAGE_ACCOUNT_NAME}.dfs.core.windows.net",
-        f"https://login.microsoftonline.com/{AZURE_TENANT_ID}/oauth2/token",
-    )
 
 
 def load_json_from_blob(container: str, file_path: str) -> dict:
